Typecasted into a byte for safety

+++ include/drivers/keyboard.h
@@ -6,10 +6,10 @@
 #include "kernel/io.h"
 
 
-#define ARROW_KEY_UP 0x80
-#define ARROW_KEY_DOWN 0x81
-#define ARROW_KEY_LEFT 0x82
-#define ARROW_KEY_RIGHT 0x83
+#define ARROW_KEY_UP (uint8_t)0x80
+#define ARROW_KEY_DOWN (uint8_t)0x81
+#define ARROW_KEY_LEFT (uint8_t)0x82
+#define ARROW_KEY_RIGHT (uint8_t)0x83
 
 
 void keyboard_init(void);
